# Conflicts:
#   webapp/src/main/java/io/zeta/metaspace/web/service/sourceinfo/SourceInfoDatabaseService.java

+++ intg/src/main/java/io/zeta/metaspace/model/dto/api/ApiTestInfoVO.java
@@ -19,7 +19,7 @@ public class ApiTestInfoVO {
     @NotNull
     String version;
     @DefaultValue("1")
-    Long pageNum;
+    Integer pageNum;
     @DefaultValue("10")
-    Long pageSize;
+    Integer pageSize;
 }
